feat/ Record Azure region when adding API key (#835)
Adds new `region` field in the APIKey model. Because Azure OpenAI
Service resources cannot be migrated, we can capture the region at the
time of API key validation.

The region is returned in the `x-ms-region` header. We can use this
region name to gate features or models based on availability.