feat(voice): Support for prompt saving + current time (#855)
Adds support for saving the assistant prompt for Voice mode chats when
creating a new thread.

Adds support for providing the current time to the assistant in Voice
mode.

Adds support for an extra "timezone" field in the Threads table, which
serves as a fallback when subsequent message requests do not contain
timezone information. This is especially useful in Voice mode, where no
data is passed along in the initial handshake.