fix/ Study Website and Server (#950)
## Study Dashboard
### New Features
- Adds new alert that pre-assessment completion rate targets will be
available on the dashboard starting August 18.
- Adds new alert that lists of students that have completed the
pre-assessment will be available on the dashboard starting August 22.
### Updates & Improvements
- When requesting the instructor classes, information related to
accepted courses (e.g. randomization status, PingPong URL) are returned
only when the course is accepted into the study to avoid unintended
information disclosure while the course is not formally accepted into
the study.
### Resolved Issues
- Fixed: Instructors may be unable to login to the Study website using
the login link emailed to them.