fix/ Clean up anonymous link file permissions after thread creation (#1148)
## Anonymous Sessions
### Resolved Issues
- Fixed: The permission to delete a user file for the associated meta
user of the shared link is not removed after the thread and Anonymous
Session token are created. The associated meta user of the shared link
does not have permissions to view the file.

## Threads
### Resolved Issues
- Fixed: When creating a thread, the server might not raise an error
because disconnecting user records from the thread database record could
fail.

## Internal
### Resolved Issues
- Fixed: The mock authentication server may fail OpenFGA's `read` checks
because responses do not include a `timestamp`.

Closes #925