feat/ Support for MCP Servers (#1156)
## Assistants
### New Features
- Use remote MCP servers with Next-Gen Assistants to give models new
capabilities. In addition to tools you make available to the model with
built in tools, you can give models new capabilities using remote MCP
servers. These tools give the model the ability to connect to and
control external services when needed to respond to a user's message.
- Remote MCP servers can be any server on the public Internet that
implements a remote [Model Context
Protocol](https://modelcontextprotocol.io/introduction) (MCP) server.
- PingPong supports remote MCP servers that use no authentication,
authentication with an OAuth access token or API key token you provide,
or any number of custom headers. Access tokens need to be generated
outside of PingPong and provided when setting up a remote MCP server in
Assistant settings.

## Threads
### New Features
- When using Next-Gen Assistants, MCP Server calls are visible in the
interface, both when streaming a response and when accessing a thread
later.